Extract catalog MCP tools
